+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/arguments/ShopkeeperFilter.java
@@ -7,11 +7,13 @@ import org.checkerframework.checker.nullness.qual.Nullable;
 import com.nisovin.shopkeepers.api.shopkeeper.Shopkeeper;
 import com.nisovin.shopkeepers.api.shopkeeper.admin.AdminShopkeeper;
 import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopkeeper;
+import com.nisovin.shopkeepers.api.ui.UIType;
 import com.nisovin.shopkeepers.commands.lib.CommandInput;
 import com.nisovin.shopkeepers.commands.lib.argument.CommandArgument;
 import com.nisovin.shopkeepers.commands.lib.argument.filter.ArgumentFilter;
 import com.nisovin.shopkeepers.commands.lib.context.CommandContextView;
 import com.nisovin.shopkeepers.lang.Messages;
+import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
 import com.nisovin.shopkeepers.text.Text;
 import com.nisovin.shopkeepers.util.java.Validate;
 
@@ -66,6 +68,36 @@ public final class ShopkeeperFilter {
 			return text;
 		}
 	};
+
+	public static ArgumentFilter<@Nullable Shopkeeper> withAccess(UIType uiType) {
+		return new ArgumentFilter<@Nullable Shopkeeper>() {
+			@Override
+			public boolean test(
+					CommandInput input,
+					CommandContextView context,
+					@Nullable Shopkeeper shopkeeper
+			) {
+				if (shopkeeper == null) return false;
+
+				// If the sender is not a player, it requires the bypass permission (usually the
+				// case for the console and block command senders):
+				return ((AbstractShopkeeper) shopkeeper).canEdit(input.getSender(), true);
+			}
+
+			@Override
+			public Text getInvalidArgumentErrorMsg(
+					CommandArgument<?> argument,
+					String argumentInput,
+					@Nullable Shopkeeper value
+			) {
+				Validate.notNull(argumentInput, "argumentInput is null");
+				Text text = Messages.commandShopkeeperArgumentNoAccess;
+				text.setPlaceholderArguments(argument.getDefaultErrorMsgArgs());
+				text.setPlaceholderArguments(Collections.singletonMap("argument", argumentInput));
+				return text;
+			}
+		};
+	}
 
 	private ShopkeeperFilter() {
 	}