+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/CommandRemove.java
@@ -7,10 +7,11 @@ import org.bukkit.entity.Player;
 
 import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
 import com.nisovin.shopkeepers.api.events.PlayerDeleteShopkeeperEvent;
-import com.nisovin.shopkeepers.api.shopkeeper.Shopkeeper;
 import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopkeeper;
+import com.nisovin.shopkeepers.api.ui.DefaultUITypes;
 import com.nisovin.shopkeepers.commands.Confirmations;
 import com.nisovin.shopkeepers.commands.arguments.ShopkeeperArgument;
+import com.nisovin.shopkeepers.commands.arguments.ShopkeeperFilter;
 import com.nisovin.shopkeepers.commands.arguments.TargetShopkeeperFallback;
 import com.nisovin.shopkeepers.commands.lib.Command;
 import com.nisovin.shopkeepers.commands.lib.CommandException;
@@ -19,6 +20,7 @@ import com.nisovin.shopkeepers.commands.lib.context.CommandContextView;
 import com.nisovin.shopkeepers.commands.util.ShopkeeperArgumentUtils.TargetShopkeeperFilter;
 import com.nisovin.shopkeepers.event.ShopkeeperEventHelper;
 import com.nisovin.shopkeepers.lang.Messages;
+import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
 import com.nisovin.shopkeepers.util.bukkit.PermissionUtils;
 import com.nisovin.shopkeepers.util.bukkit.TextUtils;
 import com.nisovin.shopkeepers.util.java.ObjectUtils;
@@ -39,8 +41,13 @@ class CommandRemove extends Command {
 		this.setDescription(Messages.commandDescriptionRemove);
 
 		// Arguments:
+		// Shopkeeper filter: Ignored for non-player command senders. Also, when deleting the shops
+		// of another player, the command only lists the shops that the executing player has editing
+		// access to as well, i.e. the executing player may require the bypass permission to see the
+		// shops owned by other players.
 		this.addArgument(new TargetShopkeeperFallback(
-				new ShopkeeperArgument(ARGUMENT_SHOPKEEPER),
+				new ShopkeeperArgument(ARGUMENT_SHOPKEEPER,
+						ShopkeeperFilter.withAccess(DefaultUITypes.EDITOR())),
 				TargetShopkeeperFilter.ANY
 		));
 	}
@@ -58,9 +65,14 @@ class CommandRemove extends Command {
 		CommandSender sender = input.getSender();
 		Player senderPlayer = ObjectUtils.castOrNull(sender, Player.class);
 
-		Shopkeeper shopkeeper = context.get(ARGUMENT_SHOPKEEPER);
+		AbstractShopkeeper shopkeeper = context.get(ARGUMENT_SHOPKEEPER);
 
-		// Permission checks:
+		// Check that the sender can edit this shop:
+		if (!shopkeeper.canEdit(sender, false)) {
+			return;
+		}
+
+		// Command permission checks:
 		if (shopkeeper instanceof PlayerShopkeeper) {
 			PlayerShopkeeper playerShop = (PlayerShopkeeper) shopkeeper;
 			if (senderPlayer != null && playerShop.isOwner(senderPlayer)) {