+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/snapshot/CommandSnapshotRestore.java
@@ -8,7 +8,9 @@ import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
 import com.nisovin.shopkeepers.api.shopkeeper.Shopkeeper;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperLoadException;
 import com.nisovin.shopkeepers.api.shopkeeper.ShopkeeperSnapshot;
+import com.nisovin.shopkeepers.api.ui.DefaultUITypes;
 import com.nisovin.shopkeepers.commands.arguments.ShopkeeperArgument;
+import com.nisovin.shopkeepers.commands.arguments.ShopkeeperFilter;
 import com.nisovin.shopkeepers.commands.arguments.TargetShopkeeperFallback;
 import com.nisovin.shopkeepers.commands.arguments.snapshot.ShopkeeperSnapshotIndexArgument;
 import com.nisovin.shopkeepers.commands.lib.Command;
@@ -19,6 +21,7 @@ import com.nisovin.shopkeepers.commands.lib.context.CommandContextView;
 import com.nisovin.shopkeepers.commands.util.ShopkeeperArgumentUtils.TargetShopkeeperFilter;
 import com.nisovin.shopkeepers.config.Settings.DerivedSettings;
 import com.nisovin.shopkeepers.lang.Messages;
+import com.nisovin.shopkeepers.shopkeeper.AbstractShopkeeper;
 import com.nisovin.shopkeepers.util.bukkit.TextUtils;
 import com.nisovin.shopkeepers.util.logging.Log;
 
@@ -38,7 +41,8 @@ class CommandSnapshotRestore extends Command {
 
 		// Arguments:
 		CommandArgument<Shopkeeper> shopkeeperArgument = new TargetShopkeeperFallback(
-				new ShopkeeperArgument(ARGUMENT_SHOPKEEPER),
+				new ShopkeeperArgument(ARGUMENT_SHOPKEEPER,
+						ShopkeeperFilter.withAccess(DefaultUITypes.EDITOR())),
 				TargetShopkeeperFilter.ANY
 		);
 		this.addArgument(shopkeeperArgument);
@@ -48,10 +52,14 @@ class CommandSnapshotRestore extends Command {
 	@Override
 	protected void execute(CommandInput input, CommandContextView context) throws CommandException {
 		CommandSender sender = input.getSender();
-		Shopkeeper shopkeeper = context.get(ARGUMENT_SHOPKEEPER);
+		AbstractShopkeeper shopkeeper = context.get(ARGUMENT_SHOPKEEPER);
 		int snapshotIndex = context.get(ARGUMENT_SNAPSHOT);
 		assert snapshotIndex >= 0 && snapshotIndex < shopkeeper.getSnapshots().size();
 		int snapshotId = snapshotIndex + 1;
+
+		if (!shopkeeper.canEdit(sender, false)) {
+			return;
+		}
 
 		ShopkeeperSnapshot snapshot = shopkeeper.getSnapshot(snapshotIndex);
 		try {